feat(Migrations): Created new tables

+++ migrations/20200224234230_create_projects.up.sql
@@ -3,7 +3,7 @@ CREATE TABLE projects (
   title varchar not null,
   code varchar not null,
   is_active boolean default false not null,
-  creator_user_id int null,
+  creator_user_id bigint null,
   created_at timestamp with time zone default current_timestamp not null,
   updated_at timestamp with time zone default current_timestamp not null
 );
